deploy/infrastructure/modules/compute/eks/lambda/requirements.txt
boto3==1.34.34
kubernetes==29.0.0
PyYAML==6.0.1 